+++ src/components/UIELements/Layout/Layout.tsx
@@ -1,12 +1,18 @@
 import {FC} from 'react';
-import {KeyboardAvoidingView, Platform, ScrollView} from 'react-native';
+import {
+  KeyboardAvoidingView,
+  Platform,
+  RefreshControl,
+  ScrollView,
+} from 'react-native';
 import {useSafeAreaInsets} from 'react-native-safe-area-context';
 import styles from './Layout.style';
 import LayoutProps from './Layout.type';
+import colors from '../../../styles/colors';
 
 const behavior = Platform.OS === 'ios' ? 'padding' : 'height';
 
-const Layout: FC<LayoutProps> = ({children, Header, style}) => {
+const Layout: FC<LayoutProps> = ({children, Header, style, onRefresh}) => {
   const {top} = useSafeAreaInsets();
 
   const HeaderComponent = () => (Header ? <Header /> : <></>);
@@ -15,7 +21,17 @@ const Layout: FC<LayoutProps> = ({children, Header, style}) => {
     <KeyboardAvoidingView
       behavior={behavior}
       style={[styles.wrapper, {paddingTop: top}]}>
-      <ScrollView contentContainerStyle={[styles.content, style]}>
+      <ScrollView
+        refreshControl={
+          <RefreshControl
+            refreshing={false}
+            onRefresh={onRefresh}
+            enabled
+            colors={[colors.secondary]}
+          />
+        }
+        nestedScrollEnabled
+        contentContainerStyle={[styles.content, style]}>
         <HeaderComponent />
         {children}
       </ScrollView>