+++ src/styles/icons.ts
@@ -1,2 +1,6 @@
 export {default as PasswordEyeIcon} from '../assets/PasswordIcon.svg';
 export {default as PasswordOpenIcon} from '../assets/PasswordOpenIcon.svg';
+export {default as BackButtonIcon} from '../assets/back.svg';
+export {default as ProfileIcon} from '../assets/Profile.svg';
+export {default as TrashIcon} from '../assets/Trash.svg';
+export {default as EditIcon} from '../assets/Edit.svg';