+++ src/types/Users.type.ts
@@ -1,0 +1,11 @@
+type UsersType = {
+  users: UserDataType[];
+};
+export type UserDataType = {
+  id: number;
+  name: string;
+  email: string;
+  adminId: number;
+};
+
+export default UsersType;