+++ CHANGELOG.md
@@ -1,3 +1,5 @@
 - SignIn [#NewFeature](https://github.com/karimansary97/Zeal/pull/new/SignIn)
 
 - Add Splash screen and SignUp Screen [#NewFeature](https://github.com/karimansary97/Zeal/pull/new/SignUp)
+
+- Add Home Screen to show users [#NewFeature](https://github.com/karimansary97/Zeal/pull/new/Home)